Hold primitive group keys in blocks too
The group keys are the other half of what a grouping stores per group, and they
grew the same way: q32 keeps 12 bytes of key next to 32 bytes of aggregate
state, all of it in allocations that were copied on every growth.

`BlockedVec` moves to `datafusion-common` so `physical-plan` can use it, and
`PrimitiveGroupValueBuilder` stores its values in it. Both comparison loops
resolve the representation once and are compiled per representation, so the
flat path is unchanged.

The flat state no longer starts at a batch: reserving one up front costs 64KB
per state vector, which `ordered_aggregate_spill.slt` showed exhausting a
600KB pool.
