+++ datafusion/common/src/utils/mod.rs
@@ -18,6 +18,7 @@
 //! This module provides the bisect function, which implements binary search.
 
 pub(crate) mod aggregate;
+pub mod blocked_vec;
 pub mod expr;
 pub mod hex;
 pub mod memory;
